+++ tests/test_inference_client.py
@@ -1,8 +1,10 @@
 from unittest.mock import Mock
 
+import numpy as np
+
 from tests import factories
 from tests.utils import assert_ndarray_eq
-from torchdemon.inference_client import InferenceClient
+from torchdemon import InferenceClient
 from torchdemon.models import InferenceRequest, InferenceResult, Signal
 
 
@@ -18,14 +20,13 @@ def test_forward() -> None:
 
     forward_arg = factories.ndarray_rand(4)
     forward_kwarg = factories.ndarray_rand(6)
-    inference_result = inference_client.forward(forward_arg, input=forward_kwarg)
+    inference_output = inference_client.forward(forward_arg, input=forward_kwarg)
 
     assert mock_connection.poll.call_count == 1
     assert mock_connection.send.call_count == 1
     assert mock_connection.recv.call_count == 1
-
-    assert inference_result.client_id == mock_inference_result.client_id
-    assert_ndarray_eq(inference_result.data[0], mock_inference_result.data[0])
+    assert isinstance(inference_output, np.ndarray)
+    assert_ndarray_eq(inference_output, mock_inference_result.data[0])
 
     sent_args, _ = mock_connection.send.call_args_list[0]
     sent_inference_request = sent_args[0]